Fix further overflow issue in VBR.
This patch fixes an additional case where under extreme conditions
some of the VBR adjustment variables can wrap.

As this is called on per frame level the extra saturation checks should
not be an issue for performance.

+++ av1/encoder/pass2_strategy.c
@@ -4267,8 +4267,9 @@ void av1_twopass_postencode_update(AV1_COMP *cpi) {
       if (rc->projected_frame_size < fast_extra_thresh) {
         p_rc->vbr_bits_off_target_fast +=
             fast_extra_thresh - rc->projected_frame_size;
-        p_rc->vbr_bits_off_target_fast = AOMMIN(p_rc->vbr_bits_off_target_fast,
-                                                (4 * rc->avg_frame_bandwidth));
+        p_rc->vbr_bits_off_target_fast =
+            AOMMIN(p_rc->vbr_bits_off_target_fast,
+                   (4 * (int64_t)rc->avg_frame_bandwidth));
       }
     }
